requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.7.0
Flask-Login==0.5.0
Flask-WTF==0.14.3
email-validator==1.1.2
python-dotenv==0.17.1
slackeventsapi==2.2.1
slack-sdk==3.5.1